==> Rmd/06-manipulation.Rmd <==
---
title: "Manipulación de datos"
subtitle: "HR Analytics: Teoría y Práctica"
author: "http://pablohaya.com/contact"
date: "09/2018"
output:
  beamer_presentation: default
  ioslides_presentation: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Importamos los datos

Empleamos la funcion `load()` para cargar la tabla que habíamos preparado en el tema anterior.

```{r message=FALSE, warning=FALSE}
library(here)
load(here("data","core_dataset_clean.RData"))
```

## Filtrar

La función `subset()` permite quedarse un subconjunto de las filas de un tabla.

```{r}
subset(df, MaritalDesc == "Single") 
```

---
De manera equivalente podemos emplear filtrar como ya vimos.

```{r}
df[df$MaritalDesc == "Single", ]
```

---

Es posible combinar más de un filtro mediante el operador lógico `&`.

```{r}
subset(df, MaritalDesc == "Single" & Age < 30)
```

---
O su versión equivalente.

```{r}
df[MaritalDesc == "Single" & Age < 30, ]
```

---

Se pueden combinar operadores lógicos para realizar condiciones más complejas. 

```{r}
subset(df, (MaritalDesc == "Single" | Age < 30) & 
            Sex == "Male")
```

## Seleccionar

La función `subset()` admite el parámetro ``select()` que devuelve una nueva tabla con las columnas seleccionadas. 

```{r}
subset(df, MaritalDesc == "Single" | Age < 30, 
       select=c(EmployeeName, MaritalDesc, Age))
```

---

Empleando la notación equivalente.

```{r}
df[df$MaritalDesc == "Single" | df$Age < 30, 
   c("EmployeeName", "MaritalDesc", "Age")]
```

---
Se puede definir un rango de columnas consecutivas.

```{r}
subset(df, MaritalDesc == "Single" | Age < 30, 
       select=c(EmployeeNumber:DOB))
```

---

Se puede indicar aquellas columnas que se quieren descartar.

```{r}
subset(df, MaritalDesc == "Single" | Age < 30, 
       select=c(-EmployeeName))
```

---

Otra opción más farragosa.

```{r}
df[df$MaritalDesc == "Single" | df$Age < 30, 
   ! colnames(df) %in% "EmployeeName"]
```

---

Cuidado que el orden de los factores aquí si altera el producto.

```{r}
colnames(df) %in% "EmployeeName"
"EmployeeName" %in% colnames(df)
```

---

Una opción más drástica es eliminar una variable por completo

```{r}
df$EmployeeName <- NULL
colnames(df)
```

## Modificar el contenido de la tabla

Se puede modificar el contenido de todo una columna sin más que asignarle un nuevo valor.

```{r}
df$Sex <- "Female"
df[,1:6]
```

---

Una operación habitual es cambiar el tipo de datos de un columna. 

```{r}
class(df$EmployeeName)
```

```{r}
df$EmployeeName <- as.character(df$EmployeeName)
```

---

Si queremos asignar un valor distinto a cada celda es preciso disponer de un vector de igual longitud que número de filas.

```{r}
df$Sex <- sample(c("Female","Male"), nrow(df), replace = TRUE)
df[,1:6]
```

---

Es posible modificar el contenido de una única celda también.

```{r}
df[1, c("EmployeeName")] <- "White, Walter"
```

## Modificar el nombre de una columna

La función `colnames()` accede al nombre de las columnas.

```{r}
colnames(df)
```

---

Reasignando un vector con nuevos nombres es posible modificarlas.

```{r}
colnames(df)[1:3] <- c("NombreEmpleado", "NumeroEmpleado", "Estado") 
colnames(df)
```


## Referencias

* [Quick-R Data Management](https://www.statmethods.net/management/index.html): resumen de cómo manipular datos con las funciones estándar de R.
