==> Rmd/07-summarization.Rmd <==
---
title: "Agregar y resumir"
subtitle: "HR Analytics: Teoría y Práctica"
author: "http://pablohaya.com/contact"
date: "09/2018"
output:
  beamer_presentation: default
  ioslides_presentation: default
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Preparación

Leemos un nuevo conjunto.

```{r message=FALSE, warning=FALSE}
library(here)
df <- read.csv(here("data", "production_staff_clean.csv"))
head(df)
```
--

y nos quedamos con una tabla con las columnas numéricas únicamente.

```{r}
df_int <- df[, c("Pay", "AbutmentsHourWk1", "AbutmentsHourWk2", 
                 "DailyErrorRate", "X90dayComplaints")]
df_int
```

## Resumiendo

`summary()` es la función más básica para obtener un resumen de las columnas.

```{r}
summary(df_int)
```

--

`colSums()` y `colMeans()` son obtienen la suma, y la media de las columnas respectivamente. Sus equivalentes `rowSums()` and `rowMeans()` actuan sobre las filas.


```{r}
colSums(df_int)
```

```{r}
colMeans(df_int)
```

---

Si queremos aplicar una función cualquiera sobre las columnas de una tabla emplearemos `sapply()`.

```{r message=FALSE, warning=FALSE}
sapply(df, class)
```

---

La sintaxis equivalente de `colMeans()` con `sapply()` sería:

```{r}
sapply(df_int, mean)
```

---

Es posible indicar que se aplique más de una función.

```{r}
sapply(df_int, function(x) c(median(x), max(x), var(x)))
```

---

y también se puede personalizar los nombres de las filas para que el resumen quede más comprensible.

```{r}
sapply(df_int, function(x) c("median"=median(x), "max"=max(x), "var"=var(x)))
```

---

Podemos definir cualquier función que se nos ocurra, como por ejemplo la siguiente función de normalización:

$\frac{x-min(x)}{max(x)-min(x)}$

```{r}
df_norm <- sapply(df_int, function(x) (x-min(x))/(max(x)-min(x)))
```

---

```{r}
summary(df_norm)
```

## Agregando

La función `aggregate()` realiza la operación indicada para cada uno de los grupos especificados por el parámetro `by`. El resultado es una tabla que resume los resultados en cada fila para cada uno de los grupos.

```{r echo=TRUE}
aggregate(df_int, by=list(df$RaceDesc), FUN=mean)
```

---

Se puede incluir más de un nivel de agregación.

```{r echo=TRUE}
aggregate(df_int$Pay, by=list(df$RaceDesc, df$Position), FUN=mean)
```

## Contando

La función `table()` permite contar el número de elementos por cada grupo. 

```{r echo=TRUE}
table(df$RaceDesc)
```

---

La función `prop.table()` aplicada sobre el resultado de `table()` devuelve el porcentaje de cada grupo. 

```{r echo=TRUE}
prop.table(table(df$RaceDesc))
```

---

Incluyendo dos columnas se realiza la tabla de contigencia correspondiente.

```{r echo=TRUE}
table(df$RaceDesc, df$Position)
```
